# zenzefi_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ZenZefi REST API provides operations for resources certificate, configuration, log, user, system",
        "version": "1.0",
        "title": "ZenZefi REST API"
    },
    "host": "localhost:61000",
    "basePath": "/",
...
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "ZenZefi Public API",
                    "ZenZefi User Management"
                ],
                "summary": "Returns a list of users.",
                "operationId": "getUsersUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/UserData"
                            }
                        }
                    }
                }
            }
        },
        "/users/currentUser": {
            "get": {
                "tags": [
                    "ZenZefi Public API",
                    "ZenZefi User Management"
                ],
                "summary": "Checks if the current user is the default user. Endpoint used in the ZenZefi UI for automated logout.",
                "operationId": "currentUserUsingGET",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/users/logout": {
            "post": {
                "tags": [
                    "ZenZefi Public API",
                    "ZenZefi User Management"
                ],
                "summary": "Log out a user from ZenZefi.",
                "operationId": "logout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/users/{ids}": {
            "delete": {
                "tags": [
                    "ZenZefi Public API",
                    "ZenZefi User Management"
                ],
                "summary": "Deletes user accounts by id. The ids must valid UUIDs.",
                "operationId": "deleteUsersUsingDELETE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "ids",
                        "in": "path",
                        "description": "Users ids, must be valid UUIDs, comma separated.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/v1/users/logout": {
            "post": {
                "tags": [
                    "ZenZefi Public API V1",
                    "ZenZefi User Management"
                ],
                "summary": "Log out a user from ZenZefi.",
                "operationId": "logoutV1",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "Locale",
                        "in": "header",
                        "description": "The language for the operation. E.g. en, de",
                        "required": true,
                        "type": "string",
                        "default": "en"
                    },
                    {
                        "name": "X-Correlation-ID",
                        "in": "header",
                        "description": "Business id which identifies the request and business case.",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "UserData": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "userName": {
                    "type": "string"
                }
            }
        },
...
    }
}